dre to current listed
